backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
stripe>=12.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.3.0
# External AI dependencies removed - using only FLAN-T5 local model
asyncio==3.4.3
psycopg2-binary>=2.9.0
pyodbc>=4.0.0
sentence-transformers>=2.2.2
requests>=2.31.0
numpy>=1.24.0
# FLAN-T5 Local Model Dependencies
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
tokenizers>=0.15.0